requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.8.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.13.3
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.2.2
frozenlist==1.3.3
h11==0.14.0
httpcore
httpx
idna>=2.8,<3.0
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
opencv-python==4.7.0.72
python-dotenv==1.0.1
python-telegram-bot>=20.0
requests==2.25.1
sniffio==1.3.1
soupsieve==2.6
tqdm==4.65.0
typing_extensions==4.12.2
urllib3==1.26.15
yarl==1.8.2
